Typescript!
Mainly doing this to learn TS on something non-trivial.
Was pretty time intensive (10h?), but with a lot of reading/learning.
It's pretty powerful, especially when you import the types for APIs (Octokit REST and GraphQL).

There's only two "@ts-ignore" and one "any" type.
I'm not sure if requiring the API signature for an entire Issue and Repository
at the lowlevel components since anything passed in will need to have *all* non-optional fields in the type.
Not a problem in practice though as long as we use the Github API exclusively with these components.

Replaced the "relative-time" dep with the far more common "date-fns" because that one had types.
It's a larger lib, but modular - so tree shaking should bring it to roughly the same size.